Ensure and-or lists separated in compound list
The previous definition of andOrList included an optional trailing
separator. That allowed two adjacent and-or lists not to be separated in
compoundList and completeLineBody.

Now andOrList does not include the trailing separator, as defined in
POSIX. The separator is parsed outside andOrList and they are combined
by (<*>). The newly added manySomeAndOrLists parses and-or list
sequences correctly.