Redefine ParserT
This commit is the core of https://github.com/magicant/flesh/issues/81.
The ParserT monad is redefined as a combination of Reader, RecordT,
PushBackT, CursorT, and ExceptT. Other relevant types and functions are
also redefined in accordance with the new ParserT.

The new definition of ParserT encapsulates the correct combination of
relevant component monads. Previously, monad combination was up to the
user and prone to errors.